Update to comply fully with Rev. 3.0 of OmniLink
Added all ExtendedObjectStatuses (mostly for fun) to ensure full code
coverage of OmniLink features. Fixed several bugs that I found along the
way, including a missing Thermostat Properties value that indicates the
capabilities of a given thermostat.

Rewrote all documentation to preserve the inner workings of the OmniLink
protocol in a nice manner for the future developers of this project.

Signed-off-by: Ethan Dye <mrtops03@gmail.com>